Qt5 build fix
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@9334 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2View/src/ov_msa/PhyTrees/MSAEditorTreeViewer.cpp
@@ -433,7 +433,7 @@ void MSAEditorTreeViewerUI::sl_sortAlignment() {
 
 void MSAEditorTreeViewerUI::highlightBranches() {
     OptionsMap settings = rectRoot->getSettings();
-    settings[BRANCH_COLOR] = Qt::black;
+    settings[BRANCH_COLOR] = static_cast<int>(Qt::black);
     if(rectRoot) {
         rectRoot->updateSettings(settings);
         rectRoot->updateChildSettings(settings);
